+++ packages/moonstone/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Fixed
+
+- `moonstone/TooltipDecorator` to hide when `onDismiss` has been invoked
+
 ### Changed
 
 - `moonstone/VideoPlayer` container changes to provide a more natural 5-way focus behavior